+++ platform/assets/js/app.ts
@@ -31,7 +31,6 @@ import {installCopyAgentGuide} from "./copy_agent_guide"
 import {installOpeningCountdown} from "./opening_countdown"
 import {installPublicTools} from "./public_tools"
 import {installRegentTokenMenu} from "./regent_token_menu"
-import {installShellMenu} from "./shell_menu"
 
 const hooks = {
   ...colocatedHooks,
@@ -67,7 +66,6 @@ holdSocketDuringCookieRotation(liveSocket.getSocket() as PinnedSocket)
 liveSocket.connect()
 installStaticMarketSearch()
 installRegentTokenMenu()
-installShellMenu()
 installOpeningCountdown()
 installCopyAgentGuide()
 installAccountAuthLazyLoader()